requirements.txt
Django==2.2.4
python-dotenv==0.13.0
libsass==0.19.2
django-compressor==2.3
django-sass-processor==0.7.3
psycopg2==2.8.3
dj-database-url==0.5.0
djangorestframework==3.10.2
drf-extensions==0.5.0
django-filter==2.2.0
Markdown==3.1.1
Pygments==2.4.2
whitenoise==4.1.3
gunicorn==19.9.0
